+++ tests/unit/transformers/test_utils.py
@@ -6,7 +6,8 @@ import pandas as pd
 import pytest
 
 from rdt.transformers.utils import (
-    _any, _max_repeat, check_nan_in_transform, flatten_column_list, strings_from_regex)
+    _any, _max_repeat, check_nan_in_transform, flatten_column_list, strings_from_regex,
+    try_convert_to_dtype)
 
 
 def test_strings_from_regex_literal():
@@ -81,8 +82,10 @@ def test_check_nan_in_transform():
     """
     # Setup
     transformed = pd.Series([0.1026, 0.1651, np.nan, 0.3116, 0.6546, 0.8541, 0.7041])
+    data_without_nans = pd.Series([0.1026, 0.1651, 0.3116, 0.6546, 0.8541, 0.7041])
 
     # Run and Assert
+    check_nan_in_transform(data_without_nans, 'float')
     expected_message = (
         'There are null values in the transformed data. The reversed '
         'transformed data will contain null values'
@@ -94,3 +97,27 @@ def test_check_nan_in_transform():
 
     with pytest.warns(UserWarning, match=expected_message_integer):
         check_nan_in_transform(transformed, 'int')
+
+
+def test_try_convert_to_dtype():
+    """Test ``try_convert_to_dtype`` method.
+
+    If the data can be converted to the specified dtype, it should be converted.
+    If the data cannot be converted, a ValueError should be raised.
+    Should allow to convert integer with NaNs to float.
+    """
+    # Setup
+    data_int_with_nan = pd.Series([1.0, 2.0, np.nan, 4.0, 5.0])
+    data_not_convertible = pd.Series(['a', 'b', 'c', 'd', 'e'])
+
+    # Run
+    output_convertibe = try_convert_to_dtype(data_int_with_nan, 'str')
+    output_int_with_nan = try_convert_to_dtype(data_int_with_nan, 'int')
+    with pytest.raises(ValueError, match="could not convert string to float: 'a'"):
+        try_convert_to_dtype(data_not_convertible, 'int')
+
+    # Assert
+    expected_data_with_nan = pd.Series([1, 2, np.nan, 4, 5])
+    expected_data_covertibe = pd.Series(['1.0', '2.0', 'nan', '4.0', '5.0'])
+    pd.testing.assert_series_equal(output_int_with_nan, expected_data_with_nan)
+    pd.testing.assert_series_equal(output_convertibe, expected_data_covertibe)
